support for scala 2.11, 2.12, 2.13
