+++ build.sbt
@@ -5,8 +5,7 @@ organization := "com.github.scredis"
 scalaVersion := "2.13.0"
 crossScalaVersions := Seq("2.11.12", "2.12.8", scalaVersion.value)
 
-Compile / scalacOptions ++= Seq("-deprecation", "-feature", "-Xfatal-warnings"/*, "-Ywarn-dead-code"*/)
-Test / scalacOptions := Seq("-deprecation", "-feature")
+scalacOptions ++= Seq("-deprecation", "-feature", "-Xfatal-warnings"/*, "-Ywarn-dead-code"*/)
 
 scalacOptions in (Compile,doc) := Seq("-no-link-warnings")
 autoAPIMappings := true
