+++ src/test/scala-2.11/scredis/util/LinkedHashSetSequencing.scala
@@ -1,0 +1,1 @@
+../../../scala-2.12/scredis/util/LinkedHashSetSequencing.scala
