+++ src/test/scala/scredis/commands/SortedSetCommandsSpec.scala
@@ -1,18 +1,16 @@
 package scredis.commands
 
-import org.scalactic.Equality
 import org.scalatest._
 import org.scalatest.concurrent._
-import org.scalatest.enablers.Sequencing
 import scredis._
 import scredis.exceptions._
 import scredis.protocol.requests.SortedSetRequests._
 import scredis.tags._
-import scredis.util.LinkedHashSet
 import scredis.util.TestUtils._
 
-import scala.collection.GenTraversable
 import scala.collection.mutable.ListBuffer
+
+import scredis.util.LinkedHashSetSequencing._
 
 
 class SortedSetCommandsSpec extends WordSpec
@@ -23,25 +21,6 @@ class SortedSetCommandsSpec extends WordSpec
   
   private val client = Client()
   private val SomeValue = "HelloWorld!虫àéç蟲"
-
-  implicit def linkedHashSetSequencing[A](implicit equality: Equality[A]): Sequencing[LinkedHashSet[A]] = new Sequencing[LinkedHashSet[A]] {
-    override def containsInOrder(sequence: LinkedHashSet[A], eles: collection.Seq[Any]): Boolean = {
-      val it = eles.iterator
-      for (e <- sequence) {
-        it.takeWhile(_ != e)
-        if (it.hasNext) it.next()
-        else return false
-      }
-      true
-    }
-
-    override def containsInOrderOnly(sequence: LinkedHashSet[A], eles: collection.Seq[Any]): Boolean = {
-      sequence.toList == eles.toList
-    }
-
-    def containsTheSameElementsInOrderAs(leftSequence: LinkedHashSet[A], rightSequence: GenTraversable[Any]): Boolean =
-      leftSequence.toList == rightSequence.toList
-  }
 
   override def beforeAll(): Unit = {
     client.hSet("HASH", "FIELD", SomeValue).!
